Assignment 3 complete

+++ Python/4.py
@@ -2,6 +2,6 @@ a = []
 s = 0
 lenlist = int(input('Enter length of list: '))
 for i in range (lenlist):
-    a.append(int(input('Enter element: ')))
+    a.append(float(input('Enter element: ')))
     s = s + a[i]
 print('Average:',int(s/lenlist))
